Refactored the Divisible trait.
